API documentation

Service interactions

Exported structures and functions

CitableImage.IIIFservice — Type

An instance of an IIIF image service.

CitableImage.url — Function

Compose URL for retrieving an image from an IIIF service.

url(img, service; ht, extension)

Arguments

url Cite2Urn for an image.

service IIIFService

ht Height of resulting image in pixels.

extension Extension for file name on server.

Markdown formatting

CitableImage.markdownImage — Function

Compose markdown wrapping an IIIF URL for retrieving an image.

markdownImage(img, service; ht, caption)

Arguments

img Cite2Urn for an image.

service IIIFService

ht Height of resulting image in pixels.

CitableImage.linkedMarkdownImage — Function

Compose markdown wrapping IIIF URL for image linked to an image citation tool installation.

linkedMarkdownImage(ict, img, service; ht, caption, extension)

Arguments

ict URL of an instance of the CiteArchitecture ImageCitationTool.

img Cite2Urn for an image.

service IIIFService

ht Height of resulting image in pixels.

caption Caption to embed in resulting linked markdown string.

Internal functions

CitableImage.pctString — Function

Given a region-of-interest string, compose percent expression for IIIF.

pctString(s)

Argument

s string with four comma-separated ROI values

Given a CITE2 URN with region-of-interest subreference, compose percent expression for IIIF.

pctString(u)

CitableImage.roiFloats — Function

Given a region-of-interest string, compute percent values, with result rounded to a given number of digits.

roiFloats(s)
roiFloats(s, digits)

Arguments

s string with four comma-separated ROI values

digits precision of resulting float